chatbortanzal/requirements.txt
groq
python-dotenv
flask
flask-cors
numpy
scikit-learn
